Build Machinekit
This is a stab at building Machinekit.  It requires a few fixes in
Machinekit to work:

- Add package release to `debian/changelog` for `3.0 (quilt)`-format
  package
- Package dep fixes to use Jessie and Trusty base packages
  - `libwebsockets-dev (>= 1.2.2)` (incorrect in DA repo)
  - `libzmq3-dev (>= 4.0.4)` (package not renamed for v. 4)

In addition to adding Machinekit package config, some changes had to
be made because there is no separate Debianization.  The scripts can't
find the package version from `debian/changelog` until the tarball is
already unpacked, so the tarball is given a temporary file name,
unpacked, and then renamed once the version is determined.

Addresses #15; fixes #16

This really needs the fix from #17 before it's suitable for use in a
Buildbot setup.